Refactoring remaining test files
